Binary header rework.
Header is now not required with the CRC calculation inside the driver (one of
the previous commits) and a fail back value for pll_init_size (this commit).
The header that is defined right now is just a skeleton for a new header that
will contain other relevant information (e.g. for sensor & camera module
probing).

Signed-off-by: Stefan Hladnik <stefan.hladnik@gmail.com>